Extracts information on comments including course, year, comments, number of thumbsups and number of thumbsdowns.

Extracts information on comments including course, year, comments, number of thumbsups and number of thumbsdowns.

comment_info(url, y = 2018)

Arguments

url

A Character value indicating the url of the webpage corresponding to an instructor.

y

A number indicating the user are interested in comments after that year.

Examples

url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
comment_info(url = url, y = 2018)
#>     course year
#> 1  STAT231 2021
#> 2  STAT231 2020
#> 3  STAT231 2019
#> 4  STAT231 2019
#> 5  STAT231 2019
#> 6  STAT588 2018
#>                                                                                                                                                                                                                                                                                                                                          comments
#> 1                                   Farzad is a fun guy who genuinely cares about his students. When our class didnt do all that well on our first exam, he took ownership of it and reevaluated how our tests are written. I had the option of watching the lectures online, but I think Farzad was much more effective and enjoyable in person.
#> 2                                                                                                                                                                                                                        Very good professor who knows a lot about the material for this class. Always happy and willing to answer any questions.
#> 3                         Farzad is by far the best professor that I have had here at Iowa State. He is incredibly caring, and always takes the time to answer students' questions. There is a lot of homework for this class, but most of it is not very difficult. If you simply show up to class and do the homework, this class is an easy A.
#> 4                                                                                                                                                                                                                                                                 Excellent professor, easy to get along with and always has time fore questions.
#> 5                                                                                                                           Farzad is a good teacher who is accessible outside of class. Hes very knowledgeable and tries to make the class humorous. You can get an A if you attend office hours, study lectures, and create a good cheat sheet.
#> 6 Stat 588 is difficult if you are not great at calculus, which I am not. Despite this, Dr. Sabzikar has EXTREMELY lenient grading practices. If you do all the work you can get a passing grade. Dr. Sabzikar explains things clearly, gives homework that is not a waste of time, and is accessible outside of class. Take this class from him!
#>   thumbsup thumbsdown
#> 1        0          0
#> 2        1          0
#> 3        1          0
#> 4        2          0
#> 5        2          0
#> 6        7          2