Sentiment Plot

A word cloud of common positive words and negative words extracted from comments, and tags from the website.

sentiment_plot(url, y = 2018, word = "Positive")

Arguments

url

A character value indicating the URL of the professor's webpage.

y

A numeric value to filter ratings after a certain year.

word

A character value indicating the user's interest in positive words, negative words, or tags.

Value

A wordcloud of common words or tags

Examples

url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
sentiment_plot(url = url, y = 2018, word = "Positive")
#> Joining, by = "word"